add bursty and multi sampledLerp classes (working)
also fixed the carrierFreq cpp method (was 2 * 2 * pi incorrectly). unit
tests in another repository show that the values are all consistent with
the original pythonic methods. speed-wise, it seems like there's a huge
jump in numpy speed when using threadpoolexecutor on the last layer ie
the burstyMulti layer. need to investigate and replicate the threading
to achieve the speedup