complete dotTonesScaling kernel and associated benchmark
we revert to using complex<float> in interrim calculations as doubles
were too slow. with complex<float>, the wall clock/cuda event timing is
about twice as fast as the cupy CZT, but this is mainly due to the fact
that the cupy CZT incurs alot of overhead between GPU kernels. the pure
GPU kernels alone are wayyy faster