add localizer classes, latlon derived versions and td/tdfd mixins
these help to shrink actual script usage by wrapping the commonly used
things (and lots of docstrings too). the way it works now is to have
common factory methods for generating the grids upon instantiation, and
then everything else in run()/localize()/plot() calls. this helps with
the common formatting that occurs in the plot call for example.