complete thresholding kernels, add in to BurstDetector
this new method incorporates the length limitations together with the
slice indices. the result is faster, but the degree to which it is
faster will depend on the exact code. most of the speedup comes from the
case where the V1 signal indices had to be reshaped into the Nx2 start and end
index array.