modify benchmark_xcorrs to include shifts total
this is mainly to fix the issue now which occurs in nsys not being able
to run it (on windows 11? or some other reason? some other people have
similar problems too), so running it in ncu for now. ncu runs each
kernel multiple times, so the best way to run it is to match the batch
size with the number of shifts, so only 1 batch is run.